Add repository for ListenTracker
This can be the place to go get the most up to date listen tracker.

I have to send in the store every time I want to go get the
ListenTracker so it knows if the tracker it has needs to be updated,
which is also why it is a `mut` call. There might be a way for it to
have a constant reference to the store internally, but I haven't been
able to get it to work well with ownership yet.