Add persistence code
The persistence for the app is going to be purely file based for now,
so this puts all of that code (other than some of the actual file
reading, which will happen in here in the future) into the persistence
module. Hopefully with how the traits are defined I'd be able to swap
to something else fairly easily.