Use new persistence
Now each of the pieces don't define their own reading/writing of their
state (initial and current) and instead delegate to a writer that is
passed in. I tried to find a way to have the trait stored inside of the
struct, but I could not get it to work easily so decided to just pass
it in when doing the calls that require it.