Massive refactor of stats
This refactors a bunch of the stats code out into separate modules,
and changes how I'm storing the song count/song name. I'm now storing
it more like a tuple because once it gets serialized it is much easier
to read. I could try to implement the serializer for it to make it
nicer in both places, but this was easy to do and required less work
overall while still giving me enough definition to understand what's
going on.