Update dependencies
There's a new version of clap, so I'm swapping to that and using the
derive way of defining the args for my cli. This cleans things up a fair
amount, and with the subcommands I can actually split my args up in such
a way that it makes way more sense when running the app.

Because I'm changing the way the app is run, I'm bumping the version
up to 3.0.0 as well.