Add Normalized track play
I'm going to be adding a track play for apple music soon, and so
instead of having my command handle each one separately, I've decided
to have one that normalizes the data from each into a single type that
contains the information I need for the command/event to make it easier
to reason about. Upcoming will be the transforms to this normalized for
my specific types.