Implement processing listens in interactive
Previously you could only do this by running the app not in interactive
mode, but I want to allow this in interactive as well. So now it does
that and then immediately applies the events to the processor.

It also creates an App that holds some state and cleans up the function
that runs it pretty well.