Make interactive the only mode
Instead of having a whole bunch of different modes, I'm making the app
launch in interactive and that's the only way to run it. Since you
can now process listens from within the app as well, I think this makes
everything a bit easier, in that you don't have to worry about a bunch
of modes and running it multiple times.

It's a massive shift but I think it makes the app more fun to use.