Update how listen tracker repo works
This probably isn't actually a good name for this anymore, but it's
fine. Getting the current projection is now a synchronous call and it
doesn't require mutating because I want to be able to "apply" the event
right after it's saved. That way I can keep a better in memory version
of the listen tracker and just occasionally push it out to disk.

This will realistically only be a thing when processing listens, so I
think keeping them all in memory is fine.